requirements.txt
certifi==2018.11.29
chardet==3.0.4
docker==3.7.0
docker-pycreds==0.4.0
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
requests==2.21.0
six==1.12.0
smmap2==2.0.5
urllib3==1.24.1
websocket-client==0.54.0
